Update Filament resource structure: pipeline module
